camera: Revert "Power: Enable/Disable power collapse after camera open"
This reverts commit a6045cb9726ce7a8773aabae0dd00c568586b99f.

This was taken from sonyxperiadev where the origin commit tries
to solve the problem with showing stacktraces on kernel log.

---
media: camera_v2: Do not remove the QoS request at msm_close time

The QoS request is being used by the camera during v4l2
operations, without necessarily keeping the media0 device open:
this makes the kernel to continuously print stacktraces and
to obviously not set the QoS request during camera usage and
that can be bad for performance.
At msm_close time, just relax the QoS request.
---

This was commit was reverted on their kernel tree in combination
with this commit to solve the issue.