qseecom: Ignore clk_access_cnt set to 1.
Since commit "qseecom: decrease clk_access_cnt"
f99808af8b29902520973bfb14950d76b0512ab6
the clk_access_cnt is always decreased.
This doesn't reflect the previous behavior where
only the clock is decreased in case the value is
set equals to "2".

It seems the commit tries to catch issues
when the clk_access_cnt is larger than the expected
value. It states
"[..] When scm_call returns EBUSY and app_access_lock is released,
clk_access_cnt could be larger than 2 as a new qseecom thread
may come to increase clk_access_cnt in this situation. [..]"
but it is ignoring the old behavior in case the
clk_access_cnt==1 the the variable is untouched.

If the clk_access_cnt is decreased here to 0, it has side effects
in function __qseecom_disable_clk. There the method checks if
the clk_access_cnt is not zero and only the it shutdowns the clock correctly.
If it set already to zero it only returns without an action.